Last comments and refactor

+++ ethics4eu/src/main/java/com/ethics4eu/ethics4eu/service/TeacherPapersService.java
@@ -11,9 +11,15 @@ import org.springframework.stereotype.Component;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * Component class with the business logic for the API paths
+ */
 @Component
 public class TeacherPapersService {
 
+    /**
+     * Instances of the Repository Interfaces
+     */
     private final TeacherPapersRepository teacherPapersRepository;
     private final PapersAttributeRepository papersAttributeRepository;
 
@@ -23,6 +29,10 @@ public class TeacherPapersService {
         this.papersAttributeRepository = papersAttributeRepository;
     }
 
+    /**
+     * Return all the BiTeX references
+     * @return JSON with a list of all the references in BibTex format
+     */
     public List<String> getTeacherPapers(){
         List<TeacherPapers> teacherPapersList = teacherPapersRepository.findAll();
         List<String> papersList = new ArrayList<>();
@@ -34,25 +44,43 @@ public class TeacherPapersService {
         return papersList;
     }
 
+    /**
+     * Return BiTeX reference by name
+     * @param name of the reference
+     * @return the reference in BibTeX format if exist, error message if don't
+     */
     public List<String> getTeacherPapersByName(String name){
         List<TeacherPapers> teacherPapersList = teacherPapersRepository.getTeacherPapersByName(name);
-        List<String> papersList = new ArrayList<>();
-        if(teacherPapersList.size() == 1){
-            System.out.println("Encontramos el paper");
-            for(TeacherPapers paper:teacherPapersList){
-                papersList.add(paper.paperBibTex());
+        if(teacherPapersList.size() == 0){
+            return List.of("Register " + name + " does not exist.");
+        }else{
+            List<String> papersList = new ArrayList<>();
+            if(teacherPapersList.size() == 1){
+                for(TeacherPapers paper:teacherPapersList){
+                    papersList.add(paper.paperBibTex());
+                }
             }
+            return papersList;
         }
-        return papersList;
     }
 
-    public void deleteTeacherPapers(Long idTeacherPapers){
-        boolean exists = teacherPapersRepository.existsById(idTeacherPapers);
-        if(!exists){
-            throw new IllegalStateException(
-                    "The document " + idTeacherPapers + " does not exist.");
+    /**
+     * Delete BibTeX reference
+     * @param name of the reference
+     * @return deletion confirmation, error message if don't
+     */
+    public String deleteTeacherPapers(String name){
+        List<TeacherPapers> teacherPapersList = teacherPapersRepository.getTeacherPapersByName(name);
+        if(teacherPapersList.size() == 0){
+            return "Register " + name + " does not exist.";
+        }else{
+            Long idTeacherPapers = teacherPapersList.get(0).getId();
+            boolean exists = teacherPapersRepository.existsById(idTeacherPapers);
+            if(!exists){
+                return "The document " + idTeacherPapers + " does not exist.";
+            }
+            papersAttributeRepository.deleteById(idTeacherPapers);
+            return "Register " + name + " was deleted.";
         }
-        papersAttributeRepository.deleteById(idTeacherPapers);
     }
-
 }
